requirements.txt
flask==3.0.0
reportlab==4.0.7
icalendar==5.0.11
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
gunicorn==21.2.0